requirements.txt
anyio==4.12.0
beautifulsoup4==4.14.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.0
dnspython==1.16.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gspread==6.2.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.1
pur==7.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pymongo==3.12.0
pyparsing==3.2.5
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
ruff==0.14.8
sniffio==1.3.1
soupsieve==2.8
StrEnum==0.4.15
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
